+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 - If two or more matching assets are found, the names are now shown to the user so the configuration can be updated more easily
 - The user is informed about tools that exist in the cache but are not present in the configuration
+- New `status` command to show config/cache/install paths, tool counts, and if a newer version of tool-installer. It also warns when the cache contains more tools than the configuration
 
 ### Changed